Overrode default temperature region to fix units
The default temperature region in Cold Sweat appears to not correctly apply units and defaults to Minecraft Units instead of C or F where applicable. This change converts the units to MC manually so they apply as intended. The issue has been reported to the mod author and should be fixed soon hopefully.